debugging gps
git-svn-id: https://vcs.maemo.org/svn/omweather/trunk@1547 5aa3d57c-301e-0410-a7f2-8fec96c11689

+++ omweather/src/weather-home_common.h
@@ -162,10 +162,10 @@ typedef	struct{
     char	name[50];
     long	start;
     long	end;
-    float	minlat;
-    float	minlon;
-    float	maxlat;
-    float 	maxlon;
+    double	minlat;
+    double	minlon;
+    double	maxlat;
+    double	maxlon;
 }Region_item;
 /*******************************************************************************/
 typedef struct{
